app/dependencies/core/ecs_realtime_listener/requirements.txt
boto3
psycopg2-binary
requests
redis